Added the ability to exclude based on AppID
Being able to set an excluded app in the menu.json allows us to exclude apps that have the same name but are actually different programs. Right now if two apps have the same name they will both be included with no way to prevent one or the other. This lets us exclude by the AppID that is in the generated AvailableStartApps.json when the scripts run. For instance I installed the new "Microsoft Edge" but I was getting icons for both the new and the old in the menu this allowed me to exclude the old icon.